requirements.txt
# Core
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.31

# Machine Learning
scikit-learn==1.4.2
joblib==1.4.2

# API
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1

# Dashboard / Visualization
streamlit==1.36.0
plotly==5.22.0
matplotlib==3.8.4
seaborn==0.13.2

# NLP (for AI log analysis later)
spacy==3.7.4
nltk==3.9.1

# Utils
requests==2.32.3
tqdm==4.66.5
python-dateutil==2.9.0
